adapters/pagi-adapter-python/requirements.txt
grpcio==1.67.1
grpcio-tools==1.67.1
pydantic==2.10.4
pytest==8.3.4

# Optional (used if installed; adapter falls back if missing)
tiktoken==0.7.0

